Extract geometry serialization into separate GeometrySerde class
Serde is going to became more complicated due to the forthcomming optimizations
