+++ presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/GeometryType.java
@@ -20,7 +20,7 @@ import com.facebook.presto.spi.type.AbstractVariableWidthType;
 import com.facebook.presto.spi.type.TypeSignature;
 import io.airlift.slice.Slice;
 
-import static com.facebook.presto.geospatial.GeometryUtils.deserialize;
+import static com.facebook.presto.geospatial.GeometrySerde.deserialize;
 
 public class GeometryType
         extends AbstractVariableWidthType
